add siglip

+++ muse/layers/attention_utils.py
@@ -121,8 +121,7 @@ class EagerAttention:
                 scores = scores.masked_fill(causal_mask, -float('inf'))
             
         # Calculate attention weights
-        # [关键修复] 强制使用 FP32 进行 Softmax，以对齐 HuggingFace SigLIP 实现
-        attn_weights = F.softmax(scores, dim=-1, dtype=torch.float32).to(scores.dtype)
+        attn_weights = F.softmax(scores, dim=-1)
         
         # Apply dropout
         if attn_dropout > 0.0:
@@ -131,6 +130,7 @@ class EagerAttention:
         # Calculate output
         output = torch.matmul(attn_weights, v)
         return output
+
 
 class FlashAttention2:
     """FlashAttention2 implementation, also conforming to AttentionFunction protocol"""
